+++ jaorm-processor/src/main/java/io/github/ulisse1996/jaorm/processor/generation/GenerationType.java
@@ -8,5 +8,6 @@ public enum GenerationType {
     CACHE,
     CONVERTER,
     LISTENERS,
-    GENERATORS
+    GENERATORS,
+    TABLES
 }